Clear Composer cache when building a Docker image
Removing the composer cache in case of a Docker build will save up to
300 MB of space inside the image.

The cache won't be used after the build so it's save to remove it.

The .dockerenv file is created by Docker and it's a good way to
determine if we are running inside a Docker container.

See https://support.xrow.net/redmine/public/issues/11670